backend/requirements.txt
fastapi==0.128.5
uvicorn[standard]==0.40.0
pydantic==2.12.5
pydantic-settings==2.12.0
openai==1.51.0
python-dotenv==1.2.1
faiss-cpu==1.13.2
tiktoken==0.12.0
httpx==0.28.1
sse-starlette==2.1.3
slowapi==0.1.9
PyJWT==2.10.1
